Code clean complete (for now)
As a matter of state, I consider this implementation finished for now.
It will require slight (not much) refactor in the future, but I will leave
that after some other considerations have been made and discussed in the project.

Possible refactors:
* solve (name, pdb_data) args passing from FileIterator.
* may be, restructure functions in proper modules.

I do consider the general architecture working pretty well and I am happy
with the current overall design.