Seems working according to my expectations for Pro
Chunks now forcely consider the presence of a consecutive Proline. And, the
chunk size is accommodated so that the residue next to the last of the chunk
is not a proline. If because of severe proline repetition that cannot be
accommodated, single residues will be built based on pair identity.